Move constants to its separate folder
